error handling and comments
